Personal Projects/Valorant-Role-Predictor/requirements.txt
keras==2.8.0
numpy==1.21.5
Pillow==8.3.2
scikit-learn==0.24.2
matplotlib==3.4.3
tensorflow==2.7.0
requests==2.26.0
